+++ tests/conftest.py
@@ -3,6 +3,8 @@ import textwrap
 
 import pytest
 
+collect_ignore = ['ufunc_example']
+
 
 def _wrap_docstring_in_func(func_name, docstring):
     template = textwrap.dedent(r"""