+++ tests/ufunc_example/module1.py
@@ -1,0 +1,7 @@
+def foo():
+    '''A doctest...
+
+    >>> foo()
+    1
+    '''
+    return 1