+++ tests/ufunc_example/module2.py
@@ -1,0 +1,1 @@
+from _module2 import foo  # noqa: F401